Modify `.gitignore` so that it *doesn't* ignore HTML files.
The HTML files in the root directory of this repository don't belong in version control,
since they are built from the demonstration files in `pages` rather than from real lessons.
However, the HTML files in the root of actual lessons definitely *do* belong in the repo,
so the `.gitignore` shouldn't ignore them.