pandoc filter to convert blockquotes to divs
This filter converts all blockquote-with-attributes to
div-with-attributes.

A blockquote-with-attributes is defined as a blockquote that has a
header with attributes defined on it as it's first element, e.g.

    > # valid blockquote {.class}
    > text

This example would be converted into this markdown:

    <div class='class' id=id>
    # valid blockquote
    text
    </div>

And the equivalent html.

The class must be one of the allowed classes defined in the filter (i.e.
one of 'callout', 'objectives' or 'challenge').