04-loop.md: Fill in FIXME challenge titles
Naming the last challenge is difficult, because it's not really about
loops.  That challenge dates back to 8e77786c (Importing shell lesson
for novices, 2013-11-11), which doesn't explain the pedagogical
motivation for the challenge.  Maybe it's just a trick to keep
students from being to focused on the purported topic when crafting
their answers?  Maybe it's intentionally a throwback to more basic
shell syntax couched in the decoration of the current topic?  Maybe
it's a hint that you can do more with loops than vary arguments to a
single command?  In any event, I'd prefer removing this challenge
entirely, or reworking the content as an example (instead of a
challenge).