Correcting URLs in footer.
The original plan was to put the URL for this repository in a configuration file that authors could edit when creating actual lessons.  However, it looks like Pandoc only does one level of variable expansion, so variables inside `_includes/footer.html` (which is itself interpolated as a variable) aren't expanded.  Instead, we will tell authors to edit `_includes/footer.html` in the `README.md` in the `gh-pages` branch.