.gitignore: Remove _site
This landed (as _site/) in e0b6f49ec (Implementation of Greg Wilson's
template, 2014-10-21) and was updated to _site (for reasons I don't
understand) in 646a9b69 (Ignoring more files, 2014-10-23).  I expect
it's a relic of a .gitignore from a Jekyll-based site (Jekyll likes to
place the compiled site in a _site/ directory), but we dropped Jekyll
with da5fc640 (Reorganizing material to put page source in
sub-directory, 2014-11-09).