Update 06-find.md
An extra multiple choice question that highlights what I think would be a common next thing to try for a lot of people, searching for something with a space in the pattern. It might also be worth putting in a callout explaining how to use ' 's to search for a pattern that includes a space. But I'm not too sure what the convention is for callouts for questions.