Minor edits to 01-filedir.md
I changed "Users" to "users" throughout this file. While "Users" is what I think is displayed in real life (on a real computer), the figures in this lesson display "users" with a lowercase u. I'm very new to the computer programming world and when I first read this lesson I honestly wasn't sure if "Users" was the same as "users" because I know case sensitivity can be important. I think keeping the cases consistent throughout the lesson will prevent confusion around this issue for other newcomers.