Suggest deleting What's in a Name?
I am suggesting we delete this section (What's in a Name?) because I don't find it to be informative. I think most people know what a filename extension is. This lesson is already lengthy (took me more than 15 minutes in total to read through/code in my shell/wrap my head around topics...), so cutting out parts that are not directly relevant to the learning objectives may be worthwhile.