Tweak to Tab Completion section
Tab completion is so useful, but I think that it is not highlighted enough in this section. I have proposed some minor tweaks to illuminate the concept a bit further and clarify exactly how it is done. Again, total newbies to the shell (like myself) might not really understand what it means when the lesson says "press tab...", so I think clarifying at least one time to "press the tab key on keyboard" is useful.