Clarify Listing Directories and Files exercise
If the goal of the exercise was to test the learner's ability to list
directories and files, then the `ls` commands would need to be modified to get
the desired output. The modifications are far beyond what is covered in the
content. Instead, this exercise is really about organizing a directory using the
`mv` command. I changed the title of the exercise to Organizing Directories and
Files. And provided some additional background on what the learner needs to do.
This is in response to issue #173.