Incorporate quotes to grep commands after quote example
Using quotes in grep commands are safer than not using them and it makes it easier to distinguish between what one is searching for and what they are searching through. After introducing the quotes in the example searching for a two word phrase, we explain their use and then use quotes in the remaining examples.