Update index.html
using the && operator, right at the beginning, might be confusing. I propose that using a semicolon, the standard bash end of line character is better. && does work, and might be "safer" in many cases because it will stop processing the line when a failed command happens, but since cd with no arguments should never fail unless there are serious problems beyond the scope of the exercise, using that functionality is superfluous at this point.