index: Fix markdown formating issue
In the 'Getting read' box on the index page, the shell command was surrounded by a block of triple
tildes. This causes the html to display a subscript tilde before and after the command.
I assume that a triple backtick was desired instead, causing a <pre><code> box in HTML.

Signed-off-by: Kai Blin <kblin@biosustain.dtu.dk>