Corrects language: "ending in" > "matches"
Exercise referred to files "ending in" *.pdb; this changes both
occurrences of that phrase to the more accurate term "matching".

I like to use colloquial idiom in lessons, but this is incorrect
enough that it will mislead learners.  The term "matching" has
been introduced in the previous lesson to describe files selected
by wildcards.