Explanation added for two files
In the wild cards section you refer to two imaginary files pi.pdb and p5.pdb along with the propane.pdb file as if they are all in the same directory. So I added a small sentence making it clear that these two files are not in the molecules directory, but if we had them in our directory p?.pdb would find it for us.