Explanation about two files
In the Wild Cards section you refer to two imaginary files pi.pdb and p5.pdb along with propane.pdb as if they are all in the same directory. To avoid confusion, I added a short sentence explaining that we don't have pi.pdb and p5.pdb in the molecules directory but if we had them the ? wild card p?.pdb would find them for us.