Flow chart is not entirely representing for loop
I find the flow chart part of the [shell_script_for_loop_flow_chart](https://github.com/swcarpentry/shell-novice/blob/gh-pages/fig/shell_script_for_loop_flow_chart.svg) in 04-loop not really representing a for-loop in that it says `$filename found --> No` --> `End`? If `$filename` is not found, the shell will throw an error.

This PR changes the text to `Next $filename?` in the hope that would work better.