Update 02-filedir.md
The current listing of `ls -F`. on a physical printout makes it look like the command to be run is: ls -F.

This is incorrect as the period is intended to indicate the end of a sentence and not the special character '.' (current directory). It also may be unclear as a student might think they should be typing the following (valid) command: ls -F .

Recommendation: Remove all punctuation (periods, commas, semi-colons, etc.) after any command example for clarity in the documentation (both online and printed).