1) Change content of `animals.txt` in excercises Which Pipe? and Appending Data to the content of file data-shell/data/animals.txt
The excercises work with the smaller input just fine and there is
file data-shell/data/animals.txt with exactly the same content.
I think it is better to refer directly to that file in the
excercises.

2) Add location of a file `animals.txt` (data-shell/data/) to
excercises which use the file

Excercises Pipe Reading Comprehension, Pipe Construction, Which
Pipe?, and Appending Data use file `animals.txt` but the location
of the file is not mentioned and so it is not entirely easy for
participants to run the commands themselves. When the location is
mentioned, it is easier to go and try to run the commands.

3) Rephrase text of excercise Appending Data

I find introduction of the excercise a bit confusing and suggest
a few changes.