Replaced line:
It's pretty simple:
every alphanumeric character matches against itself.
After the pattern comes the name or names of the files we're searching in.

With:

The grep command searches through the file, looking for matches to the pattern specified.

To use it type grep, then the pattern we're searching for and finally the name of the file (or files) we're searching in.

(As I think it reads more clearly).