Update 04-pipefilter.md: correct 'Nelle’s Pipeline: Checking Files' section by replacing '1520' with '17' and '18 respectively.
The data-shell north-pacific-gyre/2012-07-03 folder contains a total of 19 files, 17 of which are .txt files. Using 'wc -l *.txt' will produce an output of 17 lines that lists these 17 files plus a line that lists the Total for a total number of line of 18. Given the lack of synch between the text mentioning '1520' lines and the real data file, I spent a good 20 minutes trying to figure out why my download did not show 1520 files. I now conclude that the number '1520' was used for dramatization purposes - but in my mind this is no justification for providing misleading information for novices.