Refer to `flags` consistently
I didn't think so, but it looks like flags technically count as
arguments to a command.
However, introducing them as flags then referring to them as options,
and then arguments is confusing.
I have reworded the lesson text for consistency, which avoids referring
to flags as arguments.
This also sidesteps the probably opinion-based question of
"Does a flag count as an argument to a command?".

Fix #587