Simpler example to mv files to a new directory
The example of moving files to the current directory could seem overly complicated to someone who is new to the shell.  I felt like a better way to solve a problem like this would be to show you can cd into the directory you misplaced the files in and using the mv command put the multiple files on one line pointing to the new directory.   So the solution for getting the files into the raw/ folder would be: 
$ cd analyzed
$ mv sucrose.dat maltose.dat ../raw

Seems much simpler this way.  I would also reinforce to students the importance of tab completion.   At the very least, it'd be good to show this method as an alternative to the solution provided.