Fixing keyboard commands and adding history examples
I fixed the highlighting of keyboard commands to comply with the style guide.
These included <kbd>Ctrl</kbd>+<kbd>R</kbd> for example, as well as Ctrl-C and
Enter in a few places.

I also added some explanation to the history lookup commands and provided an
additional mechanism of using !head:p to recall the last command staring with 'head'.