Example of loop & counter for beginners
Topic: This is useful when teaching students that have never been taught what a loop is before. It shows how the loop sees specific files and at that same time increments through a counter (value of the counter at the beginning and at the end).
Note: I thought that this could be useful when I taught bash about a month ago, feel free to incorporate, provide me with feedback to improve or reject.