Add catch for None type code block in lesson_check
There are times when the AST is malformed and does not emit a class for
the code element. We do not want the parser to crash when this happens,
but we also want to notify ourselves that the AST is malformed.

This should not result in an error because as we saw in
https://github.com/carpentries/styles/issues/543, the parser itself can
cause these malformations when the lesson itself renders well. Even
though we fixed the previous issue with an updated parser, problems
still persist:
https://github.com/swcarpentry/r-novice-gapminder/pull/696#issuecomment-796265728

I fully admit that this is a kludge.