Update 02-filedir.md
I suggest explaining that the --help option and man are environment dependent earlier, so instructors and learners know what to expect from the beginning. It caught me off guard during teaching, and I did not realize that the explanation was given a bit further down until a few days later. During class I am usually too nervous and focused on keeping the class together, so I just told them I would investigate why it was not working for me and we skipped to the next topic.