02-filedir.md: change order of key points
The proposed order implies three consistent blocks of information that are didactically more meaningful. Hopefully the new arrangement makes it easier for the novice to memorize the prescriptions and commands by recalling their navigation experience as it unfolds.
Firstly, all objects are introduced (filesystem, files, directories, directory, trees); no change.
Secondly, all commands are sorted from where you are now to where you can go, plus a closing note on command options.
Thirdly, all information on paths is shown (their syntax, the root path, then the absolute paths, then the relative ones, then the shorthand for the current directory, and that for the directory above).
It might still be discussed whether the correct block order is 1-3-2.