01-intro-BH-Contribution.md
Kia Ora Team,

Tena koutou me nga manaakitanga !

Here is my contribution to the carpentry training. Based on my experience in workshops, I think we should explain the basics of shell, which begins with "pwd" (directory) and moves to "cd" (navigating a directory) and then "ls" ( listing items within a directory). That was easier to understand for students. Hope this does not conflict with the team. Looking forward to receiving a positive response 

Cheers

Bennett Henzeler