Update 02-filedir.md
I noticed that later there are references to the filesystem tree with no explanation of why we should think of this as a tree. There are also references to the root directory. I added a very brief explanation that will hopefully take out part of the mystery for people new to the shell.